Bug 422366 - use target file to specify repositories and exact version of non-platform prereqs (includes some changes to "production" scripts, required to test this in local test machine ... they all match was is in master, and mentioned in other releng bugs).

+++ production/master-build.sh
@@ -71,6 +71,12 @@ if [[ -z "${buildDirectory}" ]]
 then
     echo "PROGRAM ERROR: buildDirectory returned from fn-build-dir was empty"
     exit 1
+else 
+    # this should be when we first create buildDirectory
+    echo "Making buildDirectory: ${buildDirectory}"
+    mkdir -p "${buildDirectory}"
+    # it appears GID bit is not always set correctly, so we'll do so explicitly
+    chmod -c g+s "${buildDirectory}"
 fi
 export logsDirectory="${buildDirectory}/buildlogs"
 mkdir -p "${logsDirectory}"
@@ -100,12 +106,14 @@ gitCache=$( fn-git-cache "$BUILD_ROOT" "$BRANCH" )
 aggDir=$( fn-git-dir "$gitCache" "$AGGREGATOR_REPO" )
 export LOCAL_REPO="${BUILD_ROOT}"/localMavenRepo
 
-# for now, remove any existing LOCAL_REPO, and re-fetch. 
-# At some point we may reconsider this only remove once a week, 
-# or something. 
+# In production builds, we normally specify CLEAN_LOCAL, 
+# and remove any existing LOCAL_REPO, and re-fetch. 
+# But CLEAN_LOCAL can be overridden for remote builds, quick test builds, 
+# etc. 
+export CLEAN_LOCAL=${CLEAN_LOCAL:-true}
 # We "remove" by moving to backup, in case there's ever any reason to 
 # compare "what's changed". 
-if [[ -d ${LOCAL_REPO} ]]
+if [[ -d ${LOCAL_REPO} && "${CLEAN_LOCAL}" == "true" ]]
 then
     # remove existing backup, if it exists
     rm -fr ${LOCAL_REPO}.bak 2>/dev/null
@@ -164,7 +172,6 @@ fn-write-property COMMITTER_ID
 fn-write-property MVN_DEBUG
 fn-write-property MVN_QUIET
 fn-write-property SIGNING
-fn-write-property UPDATE_BRANDING
 fn-write-property REPO_AND_ACCESS
 fn-write-property MAVEN_BREE
 fn-write-property GIT_PUSH
